+++ app/src/main/java/com/br/boh_hummm/model/Motorcycle.kt
@@ -1,9 +1,9 @@
 package com.br.boh_hummm.model
 
 data class Motorcycle(
-    val mot_id: Int = 0,
+    val mot_id: Long = 0,
     val mot_brand: String?,
     val mot_type: String?,
     val mot_cylinder_capacity: Double?,
-    val mot_use_id: Int // chave estrangeira para User
+    val mot_use_id: Long // chave estrangeira para User
 )
